Extract SUNLinSolver wrapper from ARKode
