+++ include/deal.II/sundials/arkode.h
@@ -42,6 +42,7 @@
 #    include <nvector/nvector_parallel.h>
 #  endif
 #  include <deal.II/sundials/n_vector.h>
+#  include <deal.II/sundials/sunlinsol_wrapper.h>
 
 #  include <boost/signals2.hpp>
 
@@ -54,23 +55,6 @@
 
 DEAL_II_NAMESPACE_OPEN
 
-// Forward declarations
-#  if DEAL_II_SUNDIALS_VERSION_GTE(4, 0, 0)
-#    ifndef DOXYGEN
-namespace SUNDIALS
-{
-  // Forward declaration
-  template <typename VectorType>
-  struct SundialsOperator;
-
-  template <typename VectorType>
-  struct SundialsPreconditioner;
-
-  template <typename VectorType>
-  class SundialsLinearSolverWrapper;
-} // namespace SUNDIALS
-#    endif
-#  endif
 
 // Shorthand notation for ARKODE error codes.
 #  define AssertARKode(code) Assert(code >= 0, ExcARKodeError(code))
@@ -80,41 +64,6 @@ namespace SUNDIALS
  */
 namespace SUNDIALS
 {
-#  if DEAL_II_SUNDIALS_VERSION_GTE(4, 0, 0)
-  /**
-   * Type of function objects to interface with SUNDIALS linear solvers
-   *
-   * This function type encapsulates the action of solving $P^{-1}Ax=P^{-1}b$.
-   * The LinearOperator @p op encapsulates the matrix vector product $Ax$ and
-   * the LinearOperator @p prec encapsulates the application of the
-   * preconditioner $P^{-1}z$.
-   * The user can specify function objects of this type to attach custom linear
-   * solver routines to SUNDIALS. The two LinearOperators @p op and @p prec are
-   * built internally by SUNDIALS based on user settings. The parameters are
-   * interpreted as follows:
-   *
-   * @param[in] op A LinearOperator that applies the matrix vector product
-   * @param[in] prec A LinearOperator that applies the preconditioner
-   * @param[out] x The output solution vector
-   * @param[in] b The right-hand side
-   * @param[in] tol Tolerance for the iterative solver
-   *
-   * This function should return:
-   * - 0: Success
-   * - >0: Recoverable error, ARKode will reattempt the solution and call this
-   *       function again.
-   * - <0: Unrecoverable error, the computation will be aborted and an
-   *       assertion will be thrown.
-   */
-  template <typename VectorType>
-  using LinearSolveFunction =
-    std::function<int(SundialsOperator<VectorType> &      op,
-                      SundialsPreconditioner<VectorType> &prec,
-                      VectorType &                        x,
-                      const VectorType &                  b,
-                      double                              tol)>;
-#  endif
-
   /**
    * Interface to SUNDIALS additive Runge-Kutta methods (ARKode).
    *
@@ -1336,8 +1285,8 @@ namespace SUNDIALS
     GrowingVectorMemory<VectorType> mem;
 
 #  if DEAL_II_SUNDIALS_VERSION_GTE(4, 0, 0)
-    std::unique_ptr<SundialsLinearSolverWrapper<VectorType>> linear_solver;
-    std::unique_ptr<SundialsLinearSolverWrapper<VectorType>> mass_solver;
+    std::unique_ptr<internal::LinearSolverWrapper<VectorType>> linear_solver;
+    std::unique_ptr<internal::LinearSolverWrapper<VectorType>> mass_solver;
 #  endif
 
 #  ifdef DEAL_II_WITH_PETSC
@@ -1354,91 +1303,6 @@ namespace SUNDIALS
 #  endif   // DEAL_II_WITH_PETSC
   };
 
-#  if DEAL_II_SUNDIALS_VERSION_GTE(4, 0, 0)
-
-  /**
-   * A linear operator that wraps SUNDIALS functionality.
-   */
-  template <typename VectorType>
-  struct SundialsOperator
-  {
-    /**
-     * Apply this LinearOperator to @p src and store the result in @p dst.
-     */
-    void
-    vmult(VectorType &dst, const VectorType &src) const;
-
-    /**
-     * Constructor.
-     *
-     * @param A_data Data required by @p a_times_fn
-     * @param a_times_fn A function pointer to the function that computes A*v
-     */
-    SundialsOperator(void *A_data, ATimesFn a_times_fn);
-
-  private:
-    /**
-     * Data necessary to evaluate a_times_fn.
-     */
-    void *A_data;
-
-    /**
-     * Function pointer declared by SUNDIALS to evaluate the matrix vector
-     * product.
-     */
-    ATimesFn a_times_fn;
-  };
-
-
-
-  /**
-   * A linear operator that wraps preconditioner functionality as specified by
-   * SUNDIALS. The vmult() function solves the preconditioner equation $Px=b$,
-   * i.e., it computes $x=P^{-1}b$.
-   */
-  template <typename VectorType>
-  struct SundialsPreconditioner
-  {
-    /**
-     * Apply the wrapped preconditioner, i.e., solve $Px=b$ where $x$ is the
-     * @p dst vector and $b$ the @p src vector.
-     *
-     * @param dst Result vector of the preconditioner application
-     * @param src Target vector of the preconditioner application
-     */
-    void
-    vmult(VectorType &dst, const VectorType &src) const;
-
-    /**
-     * Constructor.
-     *
-     * @param P_data Data required by @p p_solve_fn
-     * @param p_solve_fn A function pointer to the function that computes A*v
-     * @param tol Tolerance, that an iterative solver should use to judge
-     *   convergence
-     */
-    SundialsPreconditioner(void *P_data, PSolveFn p_solve_fn, double tol);
-
-  private:
-    /**
-     * Data necessary to calls p_solve_fn
-     */
-    void *P_data;
-
-    /**
-     * Function pointer to a function that computes the preconditioner
-     * application.
-     */
-    PSolveFn p_solve_fn;
-
-    /**
-     * Potential tolerance to use in the internal solve of the preconditioner
-     * equation.
-     */
-    double tol;
-  };
-
-#  endif
 
   /**
    * Handle ARKode exceptions.
